Signers: provide native signer to improve devex (#272)
* Signers: provide native signer to improve devex

* add evm native signer
