+++ platforms/algorand/src/index.ts
@@ -3,5 +3,4 @@ export * from "./chain";
 export * from "./platform";
 export * from "./types";
 export * from "./unsignedTransaction";
-
-export * as testing from "./testing";
+export * from "./signer";
