--- platforms/aptos/src/testing/index.ts
@@ -1,1 +1,0 @@
-export * from "./signer";
