+++ platforms/evm/src/index.ts
@@ -3,5 +3,4 @@ export * from './unsignedTransaction';
 export * from './platform';
 export * from './types';
 export * from './chain';
-
-export * as testing from './testing';
+export * from './signer';
