+++ platforms/solana/src/index.ts
@@ -4,5 +4,4 @@ export * from './chain';
 export * from './unsignedTransaction';
 export * from './types';
 export * from './utils';
-
-export * as testing from './testing';
+export * from './signer';
